pyproject.toml
[tool.poetry]
name = "resume-normalizer"
version = "0.1.0"
description = "Backend API for resume processing, normalization, and standardized PDF generation"
authors = []
package-mode = false
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.115.14"
uvicorn = {version = "0.34.3", extras = ["standard"]}
python-multipart = "0.0.18"
pydantic = "2.13.3"
pydantic-settings = "2.11.0"
sqlalchemy = "2.0.49"
alembic = "1.16.5"
pdfplumber = "0.11.8"
python-docx = "1.2.0"
pytesseract = "0.3.13"
pdf2image = "1.17.0"
Pillow = "11.3.0"
openai = "1.109.1"
reportlab = "4.4.10"
boto3 = "1.42.96"
python-magic = "0.4.27"
python-json-logger = "3.3.0"
slowapi = "0.1.9"
eval-type-backport = "0.2.2"
psycopg2-binary = "2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "8.4.2"
pytest-cov = "6.3.0"
httpx = "0.28.1"
moto = {version = "5.1.22", extras = ["s3"]}
ruff = "0.11.13"
mypy = "1.19.1"
boto3-stubs = {version = "1.42.96", extras = ["s3"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
